requirements.txt
certifi==2020.11.8
chardet==3.0.4
dataclasses==0.6
future==0.18.2
idna==2.10
kaggle==1.5.10
numpy==1.19.4
pandas==1.1.4
Pillow==8.0.1
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2020.4
requests==2.25.0
six==1.15.0
text-unidecode==1.3
torch==1.7.0
torchvision==0.8.1
tqdm==4.54.1
typing-extensions==3.7.4.3
urllib3==1.26.2
